requirements.txt
#Flask==1.0.3  
Flask==2.0
Flask-MySQL==1.4.0  
PyMySQL==0.9.3
uWSGI==2.0.21
mysql-connector-python
cryptography
psycopg2-binary
python-dotenv